style : utils/apis 도메인별로 함수 작성으로 수정 (#72)
